requirements.txt
Flask == 0.11.1
python_ldap == 2.4.25
Requests == 2.10.0
